init progress rating
