+++ src/components/Rating/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './Rating';
