+++ src/index.tsx
@@ -4,3 +4,4 @@ export { default as Icon } from './components/Icon';
 export { default as Message } from './components/Message';
 export { default as Card } from './components/Card';
 export { default as Dropdown } from './components/Dropdown';
+export {default as StarRating} from './components/Rating';
