+++ src/styles/index.scss
@@ -11,3 +11,6 @@
 @import '../components/Dropdown/Dropdown';
 @import '../components/Input/Input';
 @import '../components/Card/Card';
+
+//Rating
+@import '../components/Rating/Rating';
